producer/requirements.txt
kafka-python
Faker
pydantic
python-schema-registry-client
